Fix Tests
Fixed JS Tests
